Add lock icon to deactivate results and change behavior of poll activation
Use a lock icon to activate/deactivate a poll in place of resuming.

If a user already has a result active for a poll, take them to that result
instead of creating a new one.

Both of these changes are done to make it easier to use MARS on multiple
devices at the same time.