Issue #17 - Changed "start poll" and "past results" workflow.
Clicking on a poll in MyPolls now takes you to the most recent result
for that poll for the logged in user. If there are no results for a
particular user, a new result is created. In both cases the result
remains locked - there is no longer any automatic activation.